+++ prim_batch/__main__.py
@@ -3,12 +3,14 @@ import argparse
 import logging
 import os
 import platform
+import re
 import shlex
 import subprocess
 import sys
 import time
 import tomllib
 from contextlib import suppress
+from fnmatch import fnmatch
 from typing import Any
 from filelock import FileLock, Timeout as LockTimeout
 from pathlib import Path
@@ -205,14 +207,23 @@ class General(HasPredefinedConfigs):
 
 class Server(HasPredefinedConfigs):
     def __init__(self, args, general: General, server_name: str):
+        def _get_args_ctrl_args():
+            if self.args.ctrl_args:
+                for s, a in self.args.ctrl_args:
+                    if (not s) or fnmatch(self.name, s):
+                        return a
+            return None
+
         server = dict_or_default(general.server_configs.get(server_name))
         super().__init__(dict_or_default(server.get(CONFIGS)))
         self.args = args
         self.general = general
+        self.name = server_name
         self.ctrl_args = shlex_split(server.get(CTRL_ARGS))
         self.sync_args = shlex_split(server.get(SYNC_ARGS))
         self.sync_args_vpn = shlex_split(server.get(SYNC_ARGS_VPN))
         self.folder_configs = dict_or_default(server.get(FOLDERS))
+        self.args_ctrl_args = shlex_split(_get_args_ctrl_args())
         self._ctrl_cmd_args = None
 
     @property
@@ -234,15 +245,15 @@ class Server(HasPredefinedConfigs):
 
     def test(self):
         test_ctrl_args = self.ctrl_cmd_args.copy()
-        if self.args.ctrl_args:
-            test_ctrl_args.extend(arg for arg in shlex_split(self.args.ctrl_args) if arg not in ['-ac', '--accept-cellular'])
+        if self.args_ctrl_args:
+            test_ctrl_args.extend(arg for arg in self.args_ctrl_args if arg not in ['-ac', '--accept-cellular'])
         exitcode, _stdout = execute('prim-ctrl', test_ctrl_args, self.args)
         return exitcode == 0
 
     def start(self, no_state: bool = False):
         start_ctrl_args = self.ctrl_cmd_args.copy()
-        if self.args.ctrl_args:
-            start_ctrl_args.extend(shlex_split(self.args.ctrl_args))
+        if self.args_ctrl_args:
+            start_ctrl_args.extend(self.args_ctrl_args)
         start_ctrl_args.extend(['-i', 'start'])
         if not no_state and "--tailscale" in self.ctrl_cmd_args:
             start_ctrl_args.extend(['-b'])
@@ -261,8 +272,8 @@ class Server(HasPredefinedConfigs):
 
     def stop(self, no_state: bool = False):
         stop_ctrl_args = self.ctrl_cmd_args.copy()
-        if self.args.ctrl_args:
-            stop_ctrl_args.extend(arg for arg in shlex_split(self.args.ctrl_args) if arg not in ['-ac', '--accept-cellular'])
+        if self.args_ctrl_args:
+            stop_ctrl_args.extend(arg for arg in self.args_ctrl_args if arg not in ['-ac', '--accept-cellular'])
         stop_ctrl_args.extend(['-i', 'stop'])
         if self.previous_state and not no_state and "--tailscale" in self.ctrl_cmd_args:
             stop_ctrl_args.extend(['-r', self.previous_state])
@@ -271,11 +282,20 @@ class Server(HasPredefinedConfigs):
 
 class Folder():
     def __init__(self, args, server: Server, folder_name: str):
+        def _get_args_sync_args():
+            if self.args.sync_args:
+                for s, f, a in self.args.sync_args:
+                    if (not s and not f) or fnmatch(self.server.name, s) and fnmatch(self.name, f):
+                        return a
+            return None
+
         self.args = args
         self.server = server
+        self.name = folder_name
         folder = dict_or_default(server.folder_configs.get(folder_name))
         self.configs = list_or_default(folder.get(CONFIGS))
         self.sync_args = shlex_split(folder.get(SYNC_ARGS))
+        self.args_sync_args = shlex_split(_get_args_sync_args())
         self._sync_cmd_args = None
 
     @property
@@ -293,8 +313,8 @@ class Folder():
             append_sync_options(self._sync_cmd_args, self.args)
             if self.args.scheduled:
                 append_if_not_in(self._sync_cmd_args, '-ss')
-            if self.args.sync_args:
-                self._sync_cmd_args.extend(shlex_split(self.args.sync_args))
+            if self.args_sync_args:
+                self._sync_cmd_args.extend(self.args_sync_args)
         return self._sync_cmd_args
 
     def sync(self):
@@ -331,10 +351,14 @@ def main():
         logging_group.add_argument('-s', '--silent', help="only errors printed", default=False, action='store_true')
         logging_group.add_argument('--debug', help="use debug level logging and add stack trace for exceptions, disables the --silent and enables the --timestamp options", default=False, action='store_true')
         ctrl_group = parser.add_argument_group('prim-ctrl')
-        ctrl_group.add_argument('--ctrl-args', metavar="ARGS", help="any prim-ctrl arguments to pass on - between quotation marks, using equal sign, like --ctrl-args='--accept-cellular'")
+        ctrl_group.add_argument('--ctrl-args', metavar="ARGS", help="any prim-ctrl arguments to pass on - between quotation marks, using equal sign, like --ctrl-args='--accept-cellular'\n"
+            "you can also specify per-server arguments using [SERVER]ARGS syntax separated by | character (SERVER can use Unix shell pattern, and can be omitted),\n"
+            "like --ctrl-args='[MyServer]--accept-cellular|[OtherServer]--debug|--silent'")
         sync_group = parser.add_argument_group('prim-sync')
         sync_group.add_argument('-d', '--dry', help="no files changed in the synchronized folder(s), only internal state gets updated and temporary files get cleaned up", default=False, action='store_true')
-        sync_group.add_argument('--sync-args', metavar="ARGS", help="any prim-sync arguments to pass on - between quotation marks, using equal sign, like --sync-args='--ignore-locks'")
+        sync_group.add_argument('--sync-args', metavar="ARGS", help="any prim-sync arguments to pass on - between quotation marks, using equal sign, like --sync-args='--ignore-locks'\n"
+            "you can also specify per-server/per-folder arguments using [SERVER,FOLDER]ARGS syntax separated by | character (SERVER and FOLDER can use Unix shell pattern, and can be omitted),\n"
+            "like --sync-args='[MyServer,MyFolder]--ignore-locks|[OtherServer,*]--debug|--dry'")
 
         args = parser.parse_args()
 
@@ -352,6 +376,14 @@ def main():
         if args.scheduled:
             print_stopped = True
             logger.info("= STARTED = %s %s", argv0, argvx)
+
+        # translate ctrl_args and sync_args into list of tuples in case of per-server/per-folder arguments
+        if args.ctrl_args:
+            regex = re.compile(r"(?:\[([^]]+)\])?([^|]+)\|?")
+            args.ctrl_args = regex.findall(args.ctrl_args)
+        if args.sync_args:
+            regex = re.compile(r"(?:\[([^,]+),([^]]+)\])?([^|]+)\|?")
+            args.sync_args = regex.findall(args.sync_args)
 
         # this testing is useful when as a scheduled task is executed after an awake and networking is not ready yet
         if args.scheduled and not test_networking(600):