+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- Optionally per-server/per-folder --ctrl-args and --sync-args arguments
 - Don't use --backup-state and --restore-state options if there is no --tailscale option is used
 - Update dependencies and project file structure